refactored package name model to models
