+++ src/main/java/be/syntra/devshop/DevshopBack/services/DTOMapperService.java
@@ -1,7 +1,7 @@
 package be.syntra.devshop.DevshopBack.services;
 
 import be.syntra.devshop.DevshopBack.entities.Product;
-import be.syntra.devshop.DevshopBack.model.ProductDTO;
+import be.syntra.devshop.DevshopBack.models.ProductDTO;
 import org.springframework.stereotype.Service;
 
 @Service
